Add Projects Page. Improve routing
