+++ packages/frontend-2/layouts/default.vue
@@ -10,7 +10,7 @@
         <Sidebar />
 
         <main class="w-full h-full overflow-y-auto simple-scrollbar pt-8 pb-16">
-          <div class="container mx-auto px-12">
+          <div class="container mx-auto px-6 md:px-12 pt-6 md:pt-0">
             <slot />
           </div>
         </main>
